Renamed nuggit.pm to Git/Nuggit and ensure scripts consistently use env in the shebang line.
The rename is consistent with Perl conventions and was done in preparation of installation of this suite for all users.

Note: Usage of "env perl" allows automatic selection of the Perl installation path for compatibility with systems that may have Perl installed to a non-standard location, or that maintain multiple Perl installations (ie: Mac users).